refactor: OrderService God Object → 3개 서비스 분리 (#14)
OrderService(610줄, 13 public 메서드, 9 의존성)를 책임별로 분리:
- OrderQueryService: 조회 전용 (getById, getList, preview 등)
- OrderCommandService: 생성·취소 (create, cancel, cancelBySystem)
- OrderPaymentService: 결제 연동 (confirm, refund, resetPaymentInProgress)

호출자 5개 수정, 테스트 3개 신규 + 5개 수정. 647개 테스트 통과.
