+++ src/main/java/com/stockmanagement/domain/order/controller/OrderController.java
@@ -13,8 +13,9 @@ import com.stockmanagement.domain.order.dto.OrderResponse;
 import com.stockmanagement.domain.order.dto.OrderSearchRequest;
 import com.stockmanagement.domain.order.dto.OrderStatusHistoryResponse;
 import com.stockmanagement.domain.order.entity.OrderStatus;
+import com.stockmanagement.domain.order.service.OrderCommandService;
 import com.stockmanagement.domain.order.service.OrderDetailService;
-import com.stockmanagement.domain.order.service.OrderService;
+import com.stockmanagement.domain.order.service.OrderQueryService;
 import com.stockmanagement.domain.user.service.UserService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
@@ -38,17 +39,6 @@ import java.util.List;
  * 주문 REST API 컨트롤러.
  *
  * <p>Base URL: {@code /api/orders}
- *
- * <pre>
- * POST /api/orders                  주문 생성 (재고 예약)              → 201 Created
- * GET  /api/orders/{id}             주문 단건 조회                     → 200 OK
- * GET  /api/orders                  주문 목록 조회 (페이징)            → 200 OK
- * POST /api/orders/{id}/cancel      주문 취소 (재고 예약 해제)         → 200 OK
- * GET  /api/orders/{id}/history     주문 상태 변경 이력 조회           → 200 OK
- * </pre>
- *
- * <p>confirm (결제 완료 처리)은 Payment 도메인에서 {@link OrderService}를 직접 호출하므로
- * 외부 API로 노출하지 않는다.
  */
 @Tag(name = "주문", description = "주문 생성 · 조회 · 취소 · 상태 이력")
 @RestController
@@ -57,7 +47,8 @@ import java.util.List;
 @Validated
 public class OrderController {
 
-    private final OrderService orderService;
+    private final OrderQueryService orderQueryService;
+    private final OrderCommandService orderCommandService;
     private final OrderDetailService orderDetailService;
     private final UserService userService;
 
@@ -67,7 +58,7 @@ public class OrderController {
             @RequestBody @Valid OrderPreviewRequest request,
             @AuthenticationPrincipal String username,
             Authentication authentication) {
-        return ApiResponse.ok(orderService.preview(request, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username))));
+        return ApiResponse.ok(orderQueryService.preview(request, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username))));
     }
 
     @Operation(summary = "주문 생성", description = "재고 예약(reserved++) 후 PENDING 주문 생성. 동일 idempotencyKey 재요청 시 기존 주문 반환.")
@@ -79,7 +70,7 @@ public class OrderController {
             @AuthenticationPrincipal String username,
             Authentication authentication) {
         Long userId = SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username));
-        return ApiResponse.ok(orderService.create(request, userId));
+        return ApiResponse.ok(orderCommandService.create(request, userId));
     }
 
     @Operation(summary = "주문 단건 조회", description = "주문 항목(items) 포함. ADMIN은 모든 주문, USER는 본인 주문만 가능.")
@@ -89,7 +80,7 @@ public class OrderController {
             @AuthenticationPrincipal String username,
             Authentication authentication) {
         boolean isAdmin = SecurityUtils.isAdmin(authentication);
-        return ApiResponse.ok(orderService.getByIdForUser(id, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username)), isAdmin));
+        return ApiResponse.ok(orderQueryService.getByIdForUser(id, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username)), isAdmin));
     }
 
     @Operation(summary = "주문 상세 통합 조회",
@@ -113,9 +104,8 @@ public class OrderController {
             @PageableDefault(size = 20, sort = "createdAt", direction = Sort.Direction.DESC)
             Pageable pageable) {
         boolean isAdmin = SecurityUtils.isAdmin(authentication);
-        // ADMIN은 request.userId 파라미터로 필터링하므로 userId 해결 불필요
         Long userId = isAdmin ? null : SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username));
-        return ApiResponse.ok(orderService.getList(userId, isAdmin, request, pageable));
+        return ApiResponse.ok(orderQueryService.getList(userId, isAdmin, request, pageable));
     }
 
     @Operation(
@@ -138,7 +128,7 @@ public class OrderController {
             @Min(1) @Max(100) @RequestParam(defaultValue = "20") int size) {
         boolean isAdmin = SecurityUtils.isAdmin(authentication);
         Long userId = SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username));
-        return ApiResponse.ok(orderService.getOrderScroll(userId, isAdmin, status, lastId, size));
+        return ApiResponse.ok(orderQueryService.getOrderScroll(userId, isAdmin, status, lastId, size));
     }
 
     @Operation(summary = "주문 취소", description = "PENDING 상태만 취소 가능. 재고 예약 해제(reserved--). ADMIN은 모든 주문 취소 가능.")
@@ -150,7 +140,7 @@ public class OrderController {
             Authentication authentication) {
         boolean isAdmin = SecurityUtils.isAdmin(authentication);
         String reason = request != null ? request.reason() : null;
-        return ApiResponse.ok(orderService.cancel(id, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username)), isAdmin, reason));
+        return ApiResponse.ok(orderCommandService.cancel(id, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username)), isAdmin, reason));
     }
 
     @Operation(summary = "주문 상태 변경 이력 조회", description = "생성·취소·확정·환불 등 모든 상태 전이를 시간순으로 반환. ADMIN은 모든 주문, USER는 본인 주문만 가능.")
@@ -160,6 +150,6 @@ public class OrderController {
             @AuthenticationPrincipal String username,
             Authentication authentication) {
         boolean isAdmin = SecurityUtils.isAdmin(authentication);
-        return ApiResponse.ok(orderService.getHistory(id, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username)), isAdmin));
+        return ApiResponse.ok(orderQueryService.getHistory(id, SecurityUtils.resolveUserId(authentication, () -> userService.resolveUserId(username)), isAdmin));
     }
 }
